+++ sketch_WSlave.ino
@@ -64,12 +64,12 @@ void software_reset()
   asm volatile ("jmp 0");
 }
 
-__attribute__((always_inline)) inline const size_t digitals_len();
-__attribute__((always_inline)) inline const size_t pulses_len();
-__attribute__((always_inline)) inline const size_t messages_len();
-const size_t digitals_len() { return ARRAYLEN(digitals); }
-const size_t pulses_len()   { return ARRAYLEN(pulses);   }
-const size_t messages_len() { return ARRAYLEN(messages); }
+__attribute__((always_inline)) inline const uint8_t digitals_len();
+__attribute__((always_inline)) inline const uint8_t pulses_len();
+__attribute__((always_inline)) inline const uint8_t messages_len();
+const uint8_t digitals_len() { return ARRAYLEN(digitals); }
+const uint8_t pulses_len()   { return ARRAYLEN(pulses);   }
+const uint8_t messages_len() { return ARRAYLEN(messages); }
